+++ package.json
@@ -17,6 +17,7 @@
     "axios": "^1.11.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.3",
+    "framer-motion": "^12.23.12",
     "react": "^19.1.0",
     "react-cookie": "^8.0.1",
     "react-dom": "^19.1.0",
