+++ src/map/constants/BottomSheetOption.tsx
@@ -1,0 +1,3 @@
+export const BOTTOM_SHEET_MIN_Y = 150; //바텀시트 최대높이
+export const BOTTOM_SHEET_MAX_Y = window.innerHeight - 300; //바텀시트 최소높이
+export const BOTTOM_SHEET_HEIGHT = window.innerHeight - BOTTOM_SHEET_MIN_Y;
