--- src/map/constants/constants.tsx
@@ -1,2 +1,0 @@
-export const BOTTOM_SHEET_MIN_Y = 150; //바텀시트 최대높이
-export const BOTTOM_SHEET_MAX_Y = window.innerHeight - 200; //바텀시트 최소높이
