+++ src/shared/api/intercepter.tsx
@@ -16,7 +16,7 @@ const api = axios.create({
 //요청 인터셉터
 api.interceptors.request.use(
   config => {
-    const access = useAuthStore.getState().accessToken;
+    const access = localStorage.getItem('AccessToken');
     if (access) {
       config.headers['Authorization'] = `Bearer ${access}`;
     }
